Version 2.3
1. Added hyphens between the parameters of the minimum / maximum damage.

2. Subtitled items CDR (+ Russian translation).
3. Fixed transitions input fields using TAB.
4. Divided Paragon points and the remaining points.
5. Made variable "Increasing DPS from stats" .
6. Done automatically imports account 5 characters (by frequency of use)
when imported into the empty list. Clearing the list is made by pressing
the right mouse key on the right side of menu with down arrow).
7. Made automatic closing block Paragon after clicking Save.
8. Fixed block CDR (save / load fields from Paragon / passive, 25%
maximum on head and right highs on other sources, an increasing of 0.5
instead of 0.1).
9. Added handling when importing passive skills Finery, Berserker Rage,
Holy Cause and shield Hellskull. Processing of Lord Commander has been
removed.
10. Minor edits interface and bug fixes.